This is in a large part a Jason Banes (jbanes) submitted patch. I took the freedom to add it to the CVS tree, along with some changes of my own (mostly due to the fact that Jason probably used SQL Server 2k and I use SQL Server 7.0, which led to some problems on my part).
These are Jason's own comments:

"I ran into the same problem as many other people with
the DatabaseMetaData (or lack thereof) implementation.
So I decided to help. This patch will fix the
implementations of:

getSchemas();
getTables();
getColumns();
supportsSchemasInTableDefinitions();

and adds support for:

getIndexes();
getPrimaryKeys();
getImportedKeys();

Unlike the previous implementation, no temporary tables
are used. The results are generated purely off of a
single SQL statement. The downside is that information
for multiple catalogs is no longer returned. However,
the JDBC specs only specify that if the catalog
variable is NULL, the catalog is not used in searching
for information. The most common implementation of this
is to only search the current database/catalog.The
upshot is that the meta queries are blazingly fast
thanks to no temp tables and don't modify the database
in any way."


git-svn-id: http://svn.code.sf.net/p/jtds/code/trunk/jtds@53 97a8069c-bbaa-4fa2-9e49-561d8e0b19ef